requirements.txt
aioredis==1.3.0
alabaster==0.7.12
amqp==2.5.1
appdirs==1.4.4
APScheduler==3.6.1
argh==0.26.2
asgiref==3.2.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.9.3
Automat==0.7.0
aws-requests-auth==0.4.2
Babel==2.7.0
bcrypt==3.1.7
billiard==3.6.1.0
black==19.10b0
bleach==3.1.5
blist==1.3.6
boto3==1.9.243
botocore==1.12.243
cached-property==1.5.1
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
channels==2.3.0
channels-redis==2.4.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configparser==4.0.2
constantly==15.1.0
croniter==0.3.30
cryptography==2.7
daphne==2.3.0
defusedxml==0.6.0
diff-match-patch==20181111
distro==1.5.0
Django==2.1.11
django-allauth==0.39.1
django-countries==5.4
django-crontab==0.7.1
django-debug-toolbar==2.0
django-elasticsearch-dsl==7.0.0
django-elasticsearch-dsl-drf==0.20.3
django-extra-fields==1.2.3
django-filter==2.2.0
django-import-export==1.2.0
django-jet==1.0.8
django-js-asset==1.2.2
django-mptt==0.10.0
django-nine==0.2.2
django-phonenumber-field==3.0.1
django-randompinfield==0.4.7
django-redis==4.10.0
django-rest-auth==0.9.5
django-twilio==0.10.0
django-uuidfield==0.5.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
docker==3.7.3
docker-compose==1.24.1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
drf-writable-nested==0.5.4
elastalert==0.2.2
elasticsearch==7.0.4
elasticsearch-dsl==7.0.0
entrypoints==0.3
envparse==0.2.0
et-xmlfile==1.0.1
exotel==0.1.5
fcm-django==0.2.21
flake8==3.7.8
future==0.17.1
googletrans==2.4.0
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hiredis==1.0.0
hpack==3.0.0
hstspreload==2020.6.16
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.7
imagesize==1.2.0
importlib-metadata==0.19
incremental==17.5.0
jdcal==1.4.1
Jinja2==2.11.1
jira==1.0.14
jmespath==0.9.4
jsonschema==2.6.0
keyring==21.2.1
kombu==4.6.4
livereload==2.6.1
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
mock==3.0.5
model-mommy==2.0.0
msgpack==0.6.2
natsort==7.0.1
numpy==1.19.0
oauthlib==3.1.0
odfpy==1.4.0
opencv-python==4.3.0.36
openpyxl==3.0.0
packaging==20.3
paramiko==2.6.0
pathspec==0.8.0
pathtools==0.1.2
pbr==5.4.3
phonenumbers==8.10.16
Pillow==6.1.0
pkginfo==1.5.0.1
port-for==0.3.1
prison==0.1.3
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pycparser==2.19
pyfcm==1.4.5
pyflakes==2.1.1
Pygments==2.6.1
PyHamcrest==1.9.0
PyJWT==1.7.1
PyNaCl==1.3.0
pyparsing==2.4.6
PyPDF2==1.26.0
pyrsistent==0.16.0
PySocks==1.7.0
PyStaticConfiguration==0.10.4
python-dateutil==2.6.1
python-decouple==3.1
python-dotenv==0.13.0
python-magic==0.4.15
python-memcached==1.59
python3-openid==3.1.0
pytz==2019.2
PyYAML==3.13
qrcode==6.1
readme-renderer==26.0
redis==3.3.8
regex==2020.5.14
requests==2.20.1
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rstcheck==3.3.1
s3transfer==0.2.1
serpy==0.3.1
six==1.12.0
sniffio==1.1.0
snowballstemmer==2.0.0
Sphinx==2.4.4
sphinx-autobuild==0.7.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.0
stomp.py==4.1.22
tablib==0.14.0
texttable==0.9.1
thehive4py==1.6.0
toml==0.10.1
tornado==6.0.4
tqdm==4.46.1
twilio==6.0.0
twine==3.2.0
Twisted==19.7.0
txaio==18.8.1
typed-ast==1.4.1
tzlocal==2.0.0
urllib3==1.24.3
vine==1.3.0
watchdog==0.10.2
webencodings==0.5.1
websocket-client==0.56.0
xlrd==1.2.0
xlwt==1.3.0
zipp==0.5.2
zope.interface==4.6.0
